requirements.txt
fastapi>=0.104.0
uvicorn>=0.23.2
gradio>=4.0.0
python-dotenv>=1.0.0
langchain>=0.1.0
langchain-openai>=0.0.2
langchain-community
faiss-cpu>=1.7.4
pypdf>=3.15.1
nest-asyncio>=1.5.8
sentence-transformers>=2.2.2